+++ changelog.md
@@ -4,6 +4,10 @@
 1. Method added 'run_as_module': Call scrapper from your own code
 2. Parameter added 'google_page_url': If you can url of the page you can go directly on that page instead of searching your 'term' on google
 
+#### Fixed
+1. Matches first 50 characters of stop_criteria review 
+
+
 
 ## 24-Nov-2023